app/requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.1
Django==3.2.6
django-extensions==3.0.9
djangorestframework==3.12.1
django-filter==2.4.0
django-cors-headers==3.5.0
